backend/pyproject.toml
[project]
name = "agentpep-backend"
version = "1.0.0"
description = "AgentPEP — Deterministic Authorization Engine for AI Agent Systems"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "motor>=3.6.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-reflection>=1.68.0",
    "jsonschema>=4.23.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "reportlab>=4.1.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
kafka = [
    "aiokafka>=0.11.0",
]
cloudwatch = [
    "boto3>=1.35.0",
]
nacl = [
    "pynacl>=1.5.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "mongomock-motor>=0.0.34",
    "hypothesis>=6.112.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["app/generated/"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
